feat(debugger): batch mode, the debug subcommand and the CLI edge
CassoCli debug --machine <name> [--disk1 <image>] [--disk2 <image>]
[--script <path>|-] [--command <line>]... [--mode applewin|monitor]
[--json] [--max-cycles <n>] [--seed <n>] [--write-disks] builds the
machine paused at power-on reset, runs the script's lines and then the
command lines through one session with every AppleWin-mode command
family attached, and exits 0, 1 (a command returned error or unknown),
2 (usage, machine or disk) or 3 (the last run ended on its budget).

- CommandLineOptions/Parser/Help: Subcommand::Debug, DebugOptions, the
  long-option table, refusals for an unknown option, a missing value, a
  surplus argument, a bad mode or seed, a missing --machine and nothing
  to run; help in every form; the debug page and the general page route.
- DebugBatchRunner: HeadlessMachineFactory with a disk card only when a
  disk is given, the DRAM seed from --seed, the budget from --max-cycles,
  a text transcript behind the mode's prompt or JSON Lines. Disk images
  are read through IFileSystem; the flush is discarded unless
  --write-disks, which writes back through the same seam. Prepare and
  Execute are separately callable so a caller can reach the machine
  between the mount and the script.
- DebugBatchSink holds notifications until the reply has printed; a
  mode change prints only as JSON, since the reply already names it.
- DebugHandlerSet attaches the eight families; FileRomSource finds
  Machines/<id>/<id>.json and the ROMs on the executable, working and
  asset directories; DebugMode is the console edge (script file or
  standard input).
- Found while generating the golden outputs and fixed: the memory bus's
  watch sink was never connected to the session's watchpoint table on a
  real machine, so BPM* could not fire outside the mock
  (IDebugTarget::SetWatchSink); and a hit during the first instruction
  of a run reported access PC $0000, because only ShouldStopBefore
  recorded it and the first instruction is never asked (OnInstruction
  records it too).
- Three warnings from earlier commits fixed (wide-to-narrow assigns,
  size_t to uint32_t).

Tests: DebugOptionsParseTests; DebugModeTests over fixture scripts and
golden text and JSON outputs under UnitTest/Fixtures/Debugger/Scripts;
CliSwitchCoverageTests and CommandLineTests extended for the debug
grammar; DebugSessionTests for the watch sink and the access PC. All
mutation-checked.

+++ CassoEmuCore/Debugger/Handlers/DataDirectiveHandlers.cpp
@@ -83,7 +83,7 @@ Word DataDirectiveHandlers::Disassemble (DebugSession & session, Word first, std
             }
         }
 
-        address += line.instruction.bytes.size();
+        address += (uint32_t) line.instruction.bytes.size();
         data.lines.push_back (line);
     }
